requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
dynaconf==3.2.10
email_validator==2.2.0
fastapi==0.115.8
flake8==7.1.2
greenlet==3.1.1
h11==0.14.0
hvac==2.3.0
idna==3.10
isort==6.0.0
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
shortuuid==1.0.13
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
win32_setctime==1.2.0
